1. Repo reorganized 2. Necessary classes added
