+++ src/app/figury/Kwadrat.java
@@ -1,0 +1,9 @@
+package app.figury;
+
+import java.awt.*;
+
+public class Kwadrat extends Figura{
+    public Kwadrat(Graphics2D buf, int del, int w, int h) {
+        super(buf, del, w, h);
+    }
+}
